requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.2.0
docx==0.2.4
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.39.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
rsa==4.9.1
setuptools==80.7.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.0
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Wave==0.0.2
